src/pyproject.toml
[tool.poetry]
name = "obcene filter"
version = "0.0.1"
description = "Example of obscene filter with demo example on django-admin and REST interface"
authors = ["zavarin-michael <zavarka@doubletapp.ai>"]

[tool.poetry.dependencies]
python = "==3.11.10"
asgi-lifespan = "==1.0.1"
httpx = "==0.23.0"
environs = "==9.5.0"
pytest = ">=7.1.3"
gunicorn = "^23.0.0"
django = "^5.1.5"
django-ninja = "^1.3.0"
django-environ = "^0.12.0"
psycopg2 = "^2.9.10"
openai = "^1.60.2"

[build-system]
requires = ["poetry-core>=1.1.13"]
build-backend = "poetry.core.masonry.api"
